requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
PyMuPDF>=1.23.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
numpy>=1.26.0
scikit-learn>=1.4.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
openai>=1.30.0
